Comparator tool moved to the hack folder
