+++ hack/shoot-comparator/README.md
@@ -1,0 +1,22 @@
+[![REUSE status](https://api.reuse.software/badge/github.com/kyma-project/infrastructure-manager)](https://api.reuse.software/info/github.com/kyma-project/infrastructure-manager)
+
+# Shoot comparator
+
+## Overview
+
+This tool compares shoot files created by KIM, and Provisioner. It was created to ensure the KIM generated exactly the same shoots as Provisioner does.
+For more details, please refer to the following issues:
+- https://github.com/kyma-project/infrastructure-manager/issues/185
+- https://github.com/kyma-project/infrastructure-manager/issues/250
+
+## Build
+```
+CGO_ENABLED=0 GOOS=darwin GOARCH=arm64 go build -o ./bin/comparator ./cmd
+```
+
+## Run
+
+In order to compare files execute the following command:
+```
+comparator files <shoot file generated by Provisioner> <shoot file generated by KIM>
+```
