+++ hack/shoot-comparator/cmd/comparator/directories.go
