+++ hack/shoot-comparator/cmd/comparator/files.go
