+++ hack/shoot-comparator/cmd/main.go
