+++ hack/shoot-comparator/internal/directories/direcotries_comparator.go
