+++ hack/shoot-comparator/internal/files/files_comparator.go
